project2/requirements.txt
Flask==2.0.1
google-cloud-speech==2.10.0
google-cloud-texttospeech==2.6.0
google-cloud-language==2.3.0

